refactor: move DisruptionView and revisionFromDisruptionForView to disruption model file
