+++ assets/src/disruptions/disruptionTable.tsx
@@ -5,7 +5,7 @@ import { Link } from "react-router-dom"
 import { formatDisruptionDate } from "./disruptions"
 import DisruptionRevision from "../models/disruptionRevision"
 import { parseDaysAndTimes } from "./time"
-import { DisruptionView } from "./viewToggle"
+import { DisruptionView } from "../models/disruption"
 import Icon from "../icons"
 import { getRouteIcon } from "./disruptionIndex"
 import { Button } from "../button"
