+++ assets/src/disruptions/viewToggle.tsx
@@ -2,14 +2,7 @@ import * as React from "react"
 import { NavLink, useHistory } from "react-router-dom"
 import queryString from "query-string"
 
-import Disruption from "../models/disruption"
-import DisruptionRevision from "../models/disruptionRevision"
-
-enum DisruptionView {
-  Draft,
-  Ready,
-  Published,
-}
+import { DisruptionView } from "../models/disruption"
 
 const useDisruptionViewParam = (): DisruptionView => {
   const { v } = queryString.parse(useHistory().location.search)
@@ -52,30 +45,4 @@ const DisruptionViewToggle = () => {
   )
 }
 
-const revisionFromDisruptionForView = (
-  disruption: Disruption,
-  view: DisruptionView
-): DisruptionRevision | undefined => {
-  switch (view) {
-    case DisruptionView.Draft: {
-      const sortedRevisions = disruption.revisions.sort((r1, r2) => {
-        return parseInt(r1.id || "", 10) - parseInt(r2.id || "", 10)
-      })
-
-      return sortedRevisions[sortedRevisions.length - 1]
-    }
-    case DisruptionView.Ready: {
-      return disruption.readyRevision
-    }
-    case DisruptionView.Published: {
-      return disruption.publishedRevision
-    }
-  }
-}
-
-export {
-  DisruptionViewToggle,
-  DisruptionView,
-  useDisruptionViewParam,
-  revisionFromDisruptionForView,
-}
+export { DisruptionViewToggle, useDisruptionViewParam }
